frameNumber and numFrames member variables are ints, not doubles
git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@7819 7c0444ca-5550-0410-9008-faecdbe78eb5

+++ src/layers/Image.hpp
@@ -159,7 +159,7 @@ protected:
    const static int RANDOM_JUMP = 1;  // const denoting jitter is a random jump
 
    //Read pvp file frame number
-   double frameNumber;
+   int frameNumber;
 
    Random * randState;
 };
