+++ src/layers/Movie.hpp
@@ -61,7 +61,7 @@ private:
    char inputfile[PV_PATH_MAX];  // current input file name
    char * movieOutputPath;  // path to output file directory for movie frames
 
-   double numFrames; //Number of frames
+   int numFrames; //Number of frames
    bool readPvpFile;
    char * fileOfFileNames;
